llvm codegen

+++ test/constexpr.c
@@ -34,8 +34,11 @@ int main() {
   ASSERT(15, ({ char x[(char)0xffffff0f]; sizeof(x); }));
   ASSERT(0x10f, ({ char x[(short)0xffff010f]; sizeof(x); }));
   ASSERT(4, ({ char x[(int)0xfffffffffff+5]; sizeof(x); }));
-  ASSERT(8, ({ char x[(int*)0+2]; sizeof(x); }));
-  ASSERT(12, ({ char x[(int*)16-1]; sizeof(x); }));
+  // Clang: error: size of array has non-integer type 'int *'
+  // ASSERT(8, ({ char x[(int*)0+2]; sizeof(x); }));
+  // ASSERT(12, ({ char x[(int*)16-1]; sizeof(x); }));
+  ASSERT(8, ({ char x[(int)((int*)0+2)]; sizeof(x); }));
+  ASSERT(12, ({ char x[(int)((int*)16-1)]; sizeof(x); }));
   ASSERT(3, ({ char x[(int*)16-(int*)4]; sizeof(x); }));
 
   ASSERT(4, ({ char x[(-1>>31)+5]; sizeof(x); }));
